General: Split location track and reference linking candidates as own components
